requirements.txt
anyio==3.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dacite==1.8.1
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.55.0
grpcio-status==1.55.0
h11==0.14.0
httpcore==0.17.1
httplib2==0.22.0
httpx==0.24.1
idna==3.4
msgpack==1.0.5
proto-plus==1.22.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
uritemplate==4.1.1
urllib3==1.26.15
